Miguelpro

+++ app/Livewire/BorrarActivo.php
@@ -19,6 +19,7 @@ class BorrarActivo extends ModalComponent
     {
         // Definir qué modelo usar basado en la vista
         $modelos = [
+            // Sucursal
             'mostraractte' => 'App\\Models\\ActivoFijo\\Activos\\ActivoTecnologia',
             'mostraractofi' => 'App\\Models\\ActivoFijo\\Activos\\ActivoOficina',
             'mostraractuni' => 'App\\Models\\ActivoFijo\\Activos\\ActivoUniforme',
@@ -34,6 +35,32 @@ class BorrarActivo extends ModalComponent
             'mostrarmob' => 'App\\Models\\ActivoFijo\\Activos\\ActivoMobiliario',
             'mostrarsou' => 'App\\Models\\ActivoFijo\\Activos\\ActivoSouvenir',
             'mostrartipoactivo' => 'App\\Models\\ActivoFijo\\Tipoactivo',
+
+            //Admin
+            'mostrarmobad' => 'App\\Models\\ActivoFijo\\Activos\\ActivoMobiliario',
+            'mostrartecad' => 'App\\Models\\ActivoFijo\\Activos\\ActivoTecnologia',
+            'mostraruniad' => 'App\\Models\\ActivoFijo\\Activos\\ActivoUniforme',
+            'mostrarsouad' => 'App\\Models\\ActivoFijo\\Activos\\ActivoSouvenir',
+            'mostrarofiad' => 'App\\Models\\ActivoFijo\\Activos\\ActivoOficina',
+            'mostrarpapead' => 'App\\Models\\ActivoFijo\\Activos\\ActivoPapeleria',
+
+            'mostrarpapead' => 'App\\Models\\ActivoFijo\\Activos\\ActivoPapeleria',
+            'mostrarnotaem' => 'App\\Models\\ActivoFijo\\Notas\\Mostrarnotaem',
+
+            'asignaciones-tec' => null,
+            'asignaciones-mob' => null,
+            'asignaciones-uni' => null,
+            'asignaciones-ofi' => null,
+            'asignaciones-pape' => null,
+            'asignaciones-sou' => null,
+
+            'asignaciones-tec-empresa' => null,
+            'asignaciones-uni-empresa' => null,
+            'asignaciones-ofi-empresa' => null,
+            'asignaciones-pape-empresa' => null,
+            'asignaciones-sou-empresa' => null,
+            'asignaciones-mob-empresa' => null,
+
         ];
 
         // Verificar si la vista tiene un modelo asignado
@@ -44,11 +71,17 @@ class BorrarActivo extends ModalComponent
         $modeloClase = $modelos[$this->vista];
 
         // Verificar si la clase del modelo existe
-        if (class_exists($modeloClase)) {
-            $activo = $modeloClase::find($this->activo_id);
-            
-            if ($activo) {
-                $activo->delete();
+        if (is_null($modeloClase)) {
+            // Generar el nombre del evento basado en la vista
+            $evento = 'deleteAsignacion' . ucfirst(str_replace('asignaciones-', '', $this->vista));
+            $this->dispatch($evento, ['rowId' => $this->activo_id]);
+        } else {
+            // Lógica existente para modelos Eloquent
+            if (class_exists($modeloClase)) {
+                $activo = $modeloClase::find($this->activo_id);
+                if ($activo) {
+                    $activo->delete();
+                }
             }
         }
