+++ app/Livewire/PortalRh/Instructor/InstructTable.php
@@ -15,6 +15,9 @@ use PowerComponents\LivewirePowerGrid\PowerGridComponent;
 use Illuminate\Support\Facades\Crypt;
 use PowerComponents\LivewirePowerGrid\Traits\WithExport; 
 use PowerComponents\LivewirePowerGrid\Components\SetUp\Exportable;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
+
 
 final class InstructTable extends PowerGridComponent
 {
@@ -31,21 +34,34 @@ final class InstructTable extends PowerGridComponent
             PowerGrid::footer()
                 ->showPerPage()
                 ->showRecordCount(),
-            PowerGrid::exportable(fileName: 'instructores-export-file') 
+            PowerGrid::exportable(fileName: 'Instructores') 
                 ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV),
         ];
     }
 
     public function datasource(): Builder
     {
-        return Instructor::query()
-        ->leftJoin('users', 'instructores.user_id', '=', 'users.id')
-        ->leftJoin('departamentos', 'instructores.departamento_id', '=', 'departamentos.id')
-        ->select([
-            'instructores.*',
-            'users.name as nombre_usuario',
-            'departamentos.nombre_departamento as departamento'
-        ]);
+        $user = Auth::user();
+    
+        $query = Instructor::query()
+            ->leftJoin('users', 'instructores.user_id', '=', 'users.id')
+            ->leftJoin('registros_patronales', 'instructores.registro_patronal_id', '=', 'registros_patronales.id')
+            ->leftJoin('departamentos', 'users.departamento_id', '=', 'departamentos.id')
+            ->leftJoin('puestos', 'users.puesto_id', '=', 'puestos.id')
+            ->select([
+                'instructores.*',
+                'users.name as nombre_usuario',
+                'registros_patronales.registro_patronal as regpatronal',
+                'departamentos.nombre_departamento as departamento',
+                'puestos.nombre_puesto as puesto',
+            ]);
+
+        // 🔹 Filtrar por departamento si es Trabajador PORTAL RH, Trabajador GLOBAL o Practicante
+        if ($user->hasRole(['Trabajador PORTAL RH', 'Trabajador GLOBAL', 'Practicante'])) {
+            $query->where('instructores.departamento_id', $user->departamento_id);
+        }
+
+        return $query;
     }
 
     public function relationSearch(): array
@@ -89,8 +105,12 @@ final class InstructTable extends PowerGridComponent
             ->add('regimen_empre')
             ->add('user_id')
             ->add('nombre_usuario')
+            ->add('registro_patronal_id')
+            ->add('regpatronal')
             ->add('departamento_id')
             ->add('departamento')
+            ->add('puesto_id')
+            ->add('puesto')
             ->add('created_at');
     }
 
@@ -98,6 +118,15 @@ final class InstructTable extends PowerGridComponent
     {
         return [
             Column::make('Id', 'id'),
+
+            Column::make('Usuario', 'nombre_usuario'),
+
+            Column::make('Departamento', 'departamento'),
+
+            Column::make('Puesto', 'puesto'),
+
+            Column::make('Registro patronal', 'regpatronal'),
+
 
             Column::make('Telefono1', 'telefono1')
                 ->sortable()
@@ -215,10 +244,6 @@ final class InstructTable extends PowerGridComponent
                 ->sortable()
                 ->searchable(),
 
-            Column::make('User id', 'nombre_usuario'),
-
-            Column::make('Departamento id', 'departamento'),
-
             Column::make('Created at', 'created_at')
                 ->sortable()
                 ->searchable(),
@@ -241,17 +266,23 @@ final class InstructTable extends PowerGridComponent
 
     public function actions(Instructor $row): array
     {
-        return [
-            Button::add('edit')
+        $actions = [];
+
+        if (Gate::allows('Editar Instructor')) {
+            $actions[] = Button::add('edit')
                 ->slot('Editar')
                 ->class('bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded')
-                ->route('editarinstructor', ['id' => Crypt::encrypt($row->id)]),
+                ->route('editarinstructor', ['id' => Crypt::encrypt($row->id)]);
+        }
 
-            Button::add('delete')
+        if (Gate::allows('Eliminar Instructor')) {
+            $actions[] = Button::add('delete')
                 ->slot('Eliminar')
                 ->class('bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded')
-                ->dispatch('confirmDelete', ['id' => $row->id]),
-        ];
+                ->dispatch('confirmDelete', ['id' => $row->id]); 
+        }
+
+        return $actions;
     }
 
     /*
